fix: undefine GetUserIdByPos

+++ inc/Config.h
@@ -32,7 +32,7 @@ bool is_use_remote_pk = false;
 enum Column{Id=0,Userid,Name,Salary};
 
 // NvmStore.h
-static char *GetUserIdByPos(uint32_t pos);
+static char *GetUserIdByPos(uint64_t pos);
  
 #define debug_db
